GCP/cloud-functions/yolo-processing/requirements.txt
opencv-python
numpy
google-cloud-storage